Deep_Learning/RAG/Streamlit_Webapp/requirements.txt
# Core dependencies
streamlit>=1.22.0
transformers>=4.29.0
sentence-transformers>=2.2.2
torch>=2.0.0
python-dotenv>=1.0.0

# RAG components
faiss-cpu>=1.7.4
langchain>=0.0.235
langchain-community>=0.0.1
langchain-huggingface>=0.0.1
python-docx
PyPDF2

# Data processing
pandas>=2.0.1
numpy>=1.24.3
pydantic>=2.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0